Update directory plugin

+++ plugins/directory/directory.plugin.zsh
@@ -1,39 +1,38 @@
 #
-# directory - Sets directory options and defines directory aliases.
+# directory - Set directory options and define directory aliases.
 #
 
-#
-# References
-#
+# References:
+# - https://github.com/sorin-ionescu/prezto/tree/master/modules/directory
 
-# https://github.com/sorin-ionescu/prezto/blob/master/modules/directory/init.zsh
+# Return if requirements are not found.
+[[ "$TERM" != 'dumb' ]] || return 1
 
-#
-# Options
-#
-
-# 16.2.1 Changing Directories
+# Set Zsh options related to directories, globbing, and I/O.
 setopt auto_pushd              # Make cd push the old directory onto the dirstack.
 setopt pushd_ignore_dups       # Don’t push multiple copies of the same directory onto the dirstack.
 setopt pushd_minus             # Exchanges meanings of +/- when navigating the dirstack.
 setopt pushd_silent            # Do not print the directory stack after pushd or popd.
 setopt pushd_to_home           # Push to home directory when no argument is given.
 setopt multios                 # Write to multiple descriptors.
-setopt glob_dots               # Don't hide dotfiles from glob patterns.
 setopt extended_glob           # Use extended globbing syntax.
+setopt glob_dots               # Don't hide dotfiles from glob patterns.
 setopt NO_clobber              # Don't overwrite files with >. Use >| to bypass.
 setopt NO_rm_star_silent       # Ask for confirmation for `rm *' or `rm path/*'
 
-#
-# Aliases
-#
-
-if ! zstyle -t ':mattmc3:zsh_custom:plugin:directory:alias' skip; then
+# Set aliases.
+if ! zstyle -t ':zsh_custom:plugin:directory:alias' skip; then
+  # directory aliases
   alias -- -='cd -'
   alias dirh='dirs -v'
-  for _idx in {1..9}; do
-    alias "$_idx"="cd +${_idx}"
-    alias -g "..$_idx"=$(printf '../%.0s' {1..$_idx})
-  done
-  unset _idx
+
+  () {
+    local i
+    for i in {1..9}; do
+      # dirstack aliases (eg: "2"="cd 2")
+      alias "$i"="cd +${i}"
+      # backref aliases (eg: "..3"="../../..")
+      alias -g "..$i"=$(printf '../%.0s' {1..$i})
+    done
+  }
 fi
